+++ docs/source/changelog.rst
@@ -17,6 +17,7 @@ This version might not be stable, but to install it use::
 Fixed
 ~~~~~
 - Escape '[' and ']' (left and right bracket).
+- Allow mappers of `~.dumps_list` to return a `~.LatexObject`.
 
 1.1.1_ - `docs <../v1.1.1/>`__ - 2016-12-10
 -------------------------------------------