+++ docs/source/changelog.rst
@@ -18,6 +18,7 @@ Fixed
 ~~~~~
 - Escape ``[`` and ``]`` (left and right bracket).
 - Allow mappers of `~.dumps_list` to return a `~.LatexObject`.
+- Section numbering default behaviour fixed
 - Setter method for `~.LatexObject.escape` property added
 
 1.1.1_ - `docs <../v1.1.1/>`__ - 2016-12-10