+++ docs/source/changelog.rst
@@ -15,6 +15,9 @@ This version might not be stable, but to install it use::
     pip install git+https://github.com/JelteF/PyLaTeX.git
 
 
+1.2.0_ - `docs <../v1.2.0/>`_
+-----------------------------
+
 Added
 ~~~~~
 - Escape flag to `.Math` container
@@ -391,7 +394,8 @@ Fixed
 - Fix package delegation with duplicate packages
 
 
-.. _Unreleased: https://github.com/JelteF/PyLaTeX/compare/v1.1.1...HEAD
+.. _Unreleased: https://github.com/JelteF/PyLaTeX/compare/v1.2.0...HEAD
+.. _1.2.0: https://github.com/JelteF/PyLaTeX/compare/v1.1.1...v1.2.0
 .. _1.1.1: https://github.com/JelteF/PyLaTeX/compare/v1.1.0...v1.1.1
 .. _1.1.0: https://github.com/JelteF/PyLaTeX/compare/v1.0.0...v1.1.0
 .. _1.0.0: https://github.com/JelteF/PyLaTeX/compare/v0.8.0...v1.0.0